CTA and footer added to every page with responsive background image
